rca_consumer/requirements.txt
flask==2.3.2
kafka-python==2.0.2
